split chain api metrics
